+++ src/data/tasks.txt
@@ -1,1 +1,2 @@
-[D][ ] d  (by 1)
+[T][ ] a
+[T][ ] b
